+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 ## v0.0.22
 
 - **Render PDF and videos within webview**
+- **Customize the protocol**
 - Fix status icon empty when in OI mode contest
 - Fix duplicated add event listener
 - Fix record list duplicated id